creation event avec date heure view account

+++ SourcesTrainingApp/src/main/java/com/sport/training/domain/service/ShoppingCartService.java
@@ -54,6 +54,6 @@ public interface ShoppingCartService {
 	 */
 	public void empty();
 
-	public Double getTotal();
+	public int getTotal();
 
 }
